refactor: API 변화에 따른 contexts actions 수정
