+++ src/axios/event/types.ts
@@ -1,10 +1,3 @@
-export interface EventListParam {
-  location: string;
-  sorting: string;
-  page: number;
-  size: number;
-}
-
 export interface GetEventParamTypes {
   eventId: string | undefined | string[];
 }
