+++ src/axios/review/getReviewList.ts
@@ -1,18 +1,13 @@
 import request from '@axios/index';
 import { ResType } from '@axios/types';
-import { ReviewList } from '@contexts/review/types';
+import { GetReviewListParam, ReviewList } from '@contexts/review/types';
 
 const getReviewList = async ({
   eventId,
   sort = 'asc',
   page,
   size = 0,
-}: {
-  eventId: string;
-  sort?: 'asc' | 'desc';
-  page: number;
-  size: number;
-}) => {
+}: GetReviewListParam) => {
   const res: ResType<ReviewList> = await request.get(
     `/events/${eventId}/?sort=createAt,${sort}&page=${page}&size=${size}`
   );
